Conflicts:
	cpp/core/dwi/tractography/mapping/voxel.h
	cpp/core/filter/connected_components.h
	cpp/core/filter/fill.h
	cpp/core/misc/bitset.cpp

+++ cpp/cmd/warpinit.cpp
@@ -66,10 +66,11 @@ void run() {
   Transform transform(header);
 
   auto func = [&transform](Image<float> &image) {
-    Eigen::Vector3d voxel_pos((float)image.index(0), (float)image.index(1), (float)image.index(2));
-    Eigen::Vector3f scanner_pos = (transform.voxel2scanner * voxel_pos).cast<float>();
+    Eigen::Vector3d voxel_pos(
+        static_cast<double>(image.index(0)), static_cast<double>(image.index(1)), static_cast<double>(image.index(2)));
+    Eigen::Vector3d scanner_pos = transform.voxel2scanner * voxel_pos;
     for (auto l = Loop(3)(image); l; ++l)
-      image.value() = scanner_pos[image.index(3)];
+      image.value() = static_cast<float>(scanner_pos[image.index(3)]);
   };
 
   ThreadedLoop("generating identity warp", warp, 0, 3).run(func, warp);
